CI: Create environment from YAML file
Allows specifying which packages to install via pip
