+++ .github/workflows/tests-workflow.yaml
@@ -51,11 +51,6 @@ jobs:
         run: |
           sudo chown -R $USER $CONDA
           ci/setup_env.sh ${{ runner.os }}
-      - name: Install Dependencies
-        run: |
-          python3 ci/list-requirements.py pyproject.toml > /tmp/conda-req.txt
-          conda install --solver=libmamba -c sebp -n sksurv-test --file /tmp/conda-req.txt
-          conda list -n sksurv-test
       - name: Configure Xcode
         run: |
           sudo xcode-select -s "/Applications/Xcode_13.2.1.app"
@@ -125,13 +120,6 @@ jobs:
         run: |
           sudo chown -R $USER $CONDA
           ci/setup_env.sh ${{ runner.os }}
-      - name: Install Dependencies
-        run: |
-          python3 ci/list-requirements.py pyproject.toml > /tmp/conda-req.txt
-          conda install --solver=libmamba -c sebp -n sksurv-test --file /tmp/conda-req.txt
-          conda install --solver=libmamba -n sksurv-test ipykernel 'matplotlib==3.5.1' nbformat 'seaborn==0.11.2' 'ipython!=8.7.0'
-          conda run -n sksurv-test pip install 'nbval>=0.10.0'
-          conda list -n sksurv-test
       - name: Configure matplotlib
         run: |
           mkdir -p ${XDG_CONFIG_HOME}/matplotlib
