+++ ci/deps/requirements.yaml.tmpl
@@ -1,0 +1,31 @@
+name: sksurv-test
+channels:
+  - sebp
+  - defaults
+dependencies:
+  - coverage
+  - cython>=0.29
+  - ecos
+  - ipykernel
+  - ipython!=8.7.0
+  - joblib
+  - matplotlib>=3.5.1,<3.6
+  - nbformat
+  - numexpr
+  - numpy={NUMPY_VERSION}
+  - osqp!=0.6.0,!=0.6.1
+  - packaging
+  - pandas={PANDAS_VERSION}
+  - pip
+  - pytest
+  - python={PYTHON_VERSION}
+  - scikit-learn={SKLEARN_VERSION}
+  - scipy>=1.3.2
+  - seaborn>=0.11.2,<0.12
+  - setuptools-scm
+  - wheel
+  - pip:
+    - build
+    - importlib-resources>=1.3
+    - nbval>=0.10.0
+    - tomli
